Allow teacher to swap unpaired students
Teacher can click an unpaired student to move them down in the list. This allows teachers to swap students to ensure optimal chatting. For example: a teacher may not want to pair up two misbehaving students.

When hovering over an unpaired student name, the student name is highlighted to show that its clickable.