Prevents keyboard from closing on mobile
If the user sends a chat message by clicking the submit button on mobile, we want the input to still be in focus. Because when the input is in focus the keyboard is displayed and the user can type out another message.   Now the input will remain in focus after sending a chat message.