Lengthened chat to fix mobile UI bug
On mobile phones, since student chatbox did not take up the whole vertical screen, there'd be lots of empty space at the bottom.   The problem with the empty space is that the autofocus on the send message input caused much of the screen to initially show much empty blank space upon starting the student chat.
To address this issue, I removed much empty space by 1) Adding a heading with a picture of the "roleplaying masks".  2) Increasing the length of the chatbox.