Moved 'student sent message' to Student's Chatbox.tsx
The 'student sent message' event was being listened to in two different files: Conversation and SendMessages.  Listening to the same event in two locations is messy and can lead to bugs where one location turns off the event for both locations.  I extracted the event to Chatbox so that the event will be listened to in a single location.